make build and run with vs 2017 and MonoGame.Framework.WindowsDX.3.6.0.1625 using nuget
note, refernces to monogram in assemblies were not touched.

usking windows with opengl there is a null refercne on startup.  content is null, cannot fix without looing at mg source.

NOte. fluid is not visible , seems to be erased at the high level calls.

update   all modules to target .net 4.5 so it can compile and work on win10 with vs 2017 <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>

remove windows.SetPosition and the openTK ref and Extensions.cs